New Equation Balancer Script
Still very incomplete. Please do not try equations with two letter elements (such as Ca or Mg) or polyatomic ions, or equations with coefficients that will be more than 1 digit when balanced (such as 12H2O). Support will be added for each case after the core logic algorithm is complete. Any contributions to solving the above problems or improving the algorithm are appreciated.